dws: add module defining cleanup routines
Problem: the coral2_dws script needs more robust routines for
ensuring that workflows are cleaned up when kubernetes is
temporarily unresponsive.

Add a new module to the flux_k8s package, `cleanup`, that will
define cleanup routines. Move the `remove_finalizer` function to
the new module.
